requirements.txt
contourpy==1.3.2
cycler==0.12.1
dotenv==0.9.9
faster-coco-eval==1.6.6
filelock==3.18.0
fonttools==4.58.1
fsspec==2025.5.1
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lightning-utilities==0.14.3
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
narwhals==1.41.0
networkx==3.5
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
plotly==6.1.2
pycocotools==2.0.9
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0
torchmetrics==1.7.2
torchvision==0.22.0
typing_extensions==4.13.2
tzdata==2025.2
